3 proposed relaxation
I propose 3 relaxations:
JUnitAssertionsShouldIncludeMessage 
A message in a JUnit assertion suffers from the same problem as a comment - it goes out of sync. A better approach would be to use fluent assertion frameworks and try to make the code read like English. 

JUnitTestContainsTooManyAsserts
This depends entirely on the type of test we are writing. Coarse grained tests, such as in-process component tests should have a decent amount of asserts. Solitary unit tests shouldn't. There is no golden rule here.

OnlyOneReturn
When I fix this, the resulting code conflicts with DataflowAnomalyAnalysis. If one of these rules has to go, I would prefer to get rid of OnlyOneReturn. This rule is a leftover from older languages where resources had to be allocated and released manually. This code practice is perfectly safe in Java.